pyproject.toml
[project]
name = "neural"
version = "0.1.0"
description = "Neural OSS platform: unified auth and providers (Kalshi signer included)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{ name = "Neural Contributors" }]
dependencies = [
  "requests>=2.31.0",
  "cryptography>=41.0.0",
  "python-dotenv>=1.0.0",
  "kalshi-python>=2.1.4",
  "websocket-client>=1.7.0",
  "simplefix>=1.0.17",
  "pandas>=2.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["neural*"]
exclude = ["secrets*"]

